+++ src/authentication/credentials.ts
@@ -3,7 +3,7 @@ import { logDebug, logError, logInfo, writeErrorFile } from "../logging/logging"
 import { StringMap } from "../types/util";
 import { encode } from "../util/base64";
 import { dedent } from "../util/dedent";
-import { LoggedError, errorMessage } from "../util/error";
+import { LoggedError, errorMessage } from "../util/errors";
 import { startInterval } from "../util/time";
 
 /**
